@@ -1,10 +1,9 @@
 package com.adityasonel.ios_textview;
 
 import android.content.Context;
-import android.graphics.Typeface;
-import android.support.v7.widget.AppCompatTextView;
+
+import androidx.appcompat.widget.AppCompatTextView;
 import android.util.AttributeSet;
-import android.view.View;
 
 public class IosTextViewBold extends AppCompatTextView {
 
