@@ -1,8 +1,8 @@
 package com.adityasonel.ios_textview;
 
 import android.content.Context;
-import android.graphics.Typeface;
-import android.support.v7.widget.AppCompatTextView;
+
+import androidx.appcompat.widget.AppCompatTextView;
 import android.util.AttributeSet;
 
 public class IosTextViewRegular extends AppCompatTextView {
